+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Terraform Provider Splunk Cloud Platform Changelog 
 
-## Version v1.0.0-beta
-* Terraform Provider v1.0.0-beta.0 testing release
+## Version v1.0.0
+* Terraform Provider inital MVP release
 * Setup provider configuration with token and username/password login support 
-* Add support for Indexes resource
+* Support for Indexes resource 
+
+## Version v1.1.0
+* Terraform Provider Phase Two release 
+* Added Indexes datasource 
+* Introduced support for HEC Token resource 
+* Introduced support for IPAllowlist resource 
+